renamed admindb to db
