+++ src/lib.rs
@@ -2,7 +2,7 @@ use std::error::Error;
 use structopt::StructOpt;
 use warp::Filter;
 
-mod admindb;
+mod db;
 pub mod error;
 
 /// opendatacapture
@@ -46,8 +46,7 @@ pub async fn run(opt: Opt) -> Result<(), Box<dyn Error>> {
         .password(opt.apiuserpassword);
     // Connect to the admin database as the default api user
     let _admindb =
-        admindb::DB::new(&dbconfig, admindb::Tableset::admin(), opt.forcereset)
-            .await?;
+        db::DB::new(&dbconfig, db::Tableset::admin(), opt.forcereset).await?;
     let routes = warp::any().map(|| "Hello, World!");
     warp::serve(routes).run(([127, 0, 0, 1], opt.apiport)).await;
     Ok(())
